[#7] refactor: 사용자 데이터 저장 로직 분리
